fix(models): P0-1 ORM critical bug fixes
- Float→Numeric for 16 monetary columns across 6 model files
  (DECIMAL precision loss prevention for billing)
- Remove broken ChunkModel alias (duplicate mapping of document_segments)
- Fix DocumentSegmentModel.remote_side (string→column object)
- Export 5 extended.py models (ABTestModel, Plugin*, ComplianceReportModel)

+++ backend/app/models/conversation.py
@@ -1,7 +1,7 @@
 """Conversation and Message models."""
 from datetime import UTC, datetime
 
-from sqlalchemy import Boolean, Column, DateTime, Float, ForeignKey, Integer, JSON, String, Text
+from sqlalchemy import Boolean, Column, DateTime, Float, ForeignKey, Integer, JSON, Numeric, String, Text
 from sqlalchemy.orm import relationship
 
 from app.models.base import Base, generate_uuid
@@ -24,7 +24,7 @@ class ConversationModel(Base):
     message_count = Column(Integer, default=0)
     total_input_tokens = Column(Integer, default=0)
     total_output_tokens = Column(Integer, default=0)
-    total_cost = Column(Float, default=0.0)
+    total_cost = Column(Numeric(10, 6), default=0.0)
     last_message_at = Column(DateTime, nullable=True, index=True)
     last_message_preview = Column(String(200), nullable=True)
     archived_at = Column(DateTime, nullable=True)
